calvin/calvin_env/requirements.txt
cloudpickle
gitpython
gym
hydra-core
hydra-colorlog
matplotlib
numba
numpy
numpy-quaternion
omegaconf
opencv-python
pandas
pybullet
scipy
rich
